Fixed issues and generated stuff
Added generate.py so that importing summarize doesn't run code or need gpt at all
Fixed bugs with the list being None
